pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "todoist-tui"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.61.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
todoist-tui = "todoist_tui.main:run"
todoist = "todoist_tui.main:run"

[tool.hatch.envs.default]
dependencies = ["textual-dev>=1.5.0"]
